requirements.txt
Flask
flask_cors
boto3
geopy
python-dotenv
python-google-places
requests